Convert menu to index based stories in folders
