+++ packages/react-components/react-menu/src/stories/Menu/MenuControlledRadioItems.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { Menu, MenuTrigger, MenuList, MenuItemRadio, MenuPopover } from '../index';
+import { Menu, MenuTrigger, MenuList, MenuItemRadio, MenuPopover, MenuProps } from '@fluentui/react-menu';
 
 import { Button } from '@fluentui/react-button';
 import {
@@ -17,14 +17,19 @@ const CutIcon = bundleIcon(CutFilled, CutRegular);
 const PasteIcon = bundleIcon(ClipboardPasteFilled, ClipboardPasteRegular);
 const EditIcon = bundleIcon(EditFilled, EditRegular);
 
-export const RadioItems = () => {
+export const ControlledRadioItems = () => {
+  const [checkedValues, setCheckedValues] = React.useState<Record<string, string[]>>({ font: ['calibri'] });
+  const onChange: MenuProps['onCheckedValueChange'] = (e, { name, checkedItems }) => {
+    setCheckedValues(s => ({ ...s, [name]: checkedItems }));
+  };
+
   return (
     <Menu>
       <MenuTrigger>
         <Button>Toggle menu</Button>
       </MenuTrigger>
       <MenuPopover>
-        <MenuList>
+        <MenuList checkedValues={checkedValues} onCheckedValueChange={onChange}>
           <MenuItemRadio icon={<CutIcon />} name="font" value="segoe">
             Segoe
           </MenuItemRadio>
@@ -38,19 +43,4 @@ export const RadioItems = () => {
       </MenuPopover>
     </Menu>
   );
-};
-
-RadioItems.parameters = {
-  docs: {
-    description: {
-      story: [
-        'A variant of `MenuItem` that handles radio like selection. The `name` and `value` props are are used',
-        'similar to [HTML checkboxes with `input`](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/radio)',
-        '',
-        '```html',
-        '<input type="radio" name="name" value="value" />',
-        '```',
-      ].join('\n'),
-    },
-  },
 };
