+++ packages/react-components/react-menu/src/stories/Menu/MenuControllingOpenAndClose.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { Menu, MenuTrigger, MenuList, MenuItem, MenuPopover, MenuProps } from '../index';
+import { Menu, MenuTrigger, MenuList, MenuItem, MenuPopover, MenuProps } from '@fluentui/react-menu';
 
 import { Button } from '@fluentui/react-button';
 import { Checkbox, CheckboxProps } from '@fluentui/react-checkbox';
