+++ packages/react-components/react-menu/src/stories/Menu/MenuCustomTrigger.stories.tsx
@@ -1,7 +1,15 @@
 import * as React from 'react';
 import { Button } from '@fluentui/react-button';
 
-import { Menu, MenuProps, MenuTrigger, MenuList, MenuItem, MenuPopover, MenuTriggerChildProps } from '../index';
+import {
+  Menu,
+  MenuProps,
+  MenuTrigger,
+  MenuList,
+  MenuItem,
+  MenuPopover,
+  MenuTriggerChildProps,
+} from '@fluentui/react-menu';
 
 const CustomMenuTrigger = React.forwardRef<HTMLButtonElement, Partial<MenuTriggerChildProps>>((props, ref) => {
   return (
