+++ packages/react-components/react-menu/src/stories/Menu/MenuGroupingItems.stories.tsx
@@ -1,5 +1,14 @@
 import * as React from 'react';
-import { Menu, MenuTrigger, MenuList, MenuItem, MenuGroup, MenuDivider, MenuGroupHeader, MenuPopover } from '../index';
+import {
+  Menu,
+  MenuTrigger,
+  MenuList,
+  MenuItem,
+  MenuGroup,
+  MenuDivider,
+  MenuGroupHeader,
+  MenuPopover,
+} from '@fluentui/react-menu';
 
 import { Button } from '@fluentui/react-button';
 import {
