+++ packages/react-components/react-menu/src/stories/Menu/MenuSelectionGroup.stories.tsx
@@ -9,7 +9,7 @@ import {
   MenuDivider,
   MenuGroupHeader,
   MenuPopover,
-} from '../index';
+} from '@fluentui/react-menu';
 
 import { Button } from '@fluentui/react-button';
 import {
