+++ packages/react-components/react-menu/src/stories/Menu/MenuVisualDividerOnly.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { Menu, MenuTrigger, MenuList, MenuItem, MenuDivider, MenuPopover } from '../index';
+import { Menu, MenuTrigger, MenuList, MenuItem, MenuDivider, MenuPopover } from '@fluentui/react-menu';
 
 import { Button } from '@fluentui/react-button';
 import {
